Convert Money Amount from float to double

+++ src/main/java/cinema/model/Projection.java
@@ -48,7 +48,7 @@ public class Projection {
 		return price;
 	}
 	// getters per ricavare le informazioni sul prezzo di queste ultime
-	public float getAmount() {
+	public double getAmount() {
 		return price.getAmount();
 	}
 	public TypeOfCurrency getCurrency() {
